requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.3.0
langchain-community==0.3.0
faiss-cpu==1.13.2
pypdf==5.0.1
python-dotenv==1.0.1
requests==2.32.3
sentence-transformers==3.1.1

